+++ src/lmgrep/lucene/monitor.clj
@@ -38,12 +38,16 @@
     (try
       (.register monitor (doto (ArrayList.) (.add mq)))
       (catch Exception e
+        (when (System/getenv "DEBUG_MODE")
+          (.printStackTrace e))
         (.println System/err (format "Failed to register query %s with exception '%s'" mq (.getMessage e)))))))
 
 (defn register-queries [^Monitor monitor monitor-queries]
   (try
     (.register monitor ^Iterable monitor-queries)
     (catch Exception e
+      (when (System/getenv "DEBUG_MODE")
+        (.printStackTrace e))
       (.println System/err (format "Failed to register queries with exception '%s'" (.getMessage e)))
       (defer-to-one-by-one-registration monitor monitor-queries))))